// src/http/models/modelsPet.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * API RMI
 * API para gerenciamento de dados de cidadãos do Rio de Janeiro, incluindo autodeclaração de informações e verificação de contato.
 * OpenAPI spec version: 1.0
 */
import type { ModelsAccreditedClinic } from './modelsAccreditedClinic'
import type { ModelsOtherTutor } from './modelsOtherTutor'

export interface ModelsPet {
  animal_nome?: string
  antirrabica_data?: string
  antirrabica_validade_data?: string
  clinica_credenciada?: ModelsAccreditedClinic
  especie_nome?: string
  fase_vida_nome?: string
  foto_url?: string
  id_animal?: number
  indicador?: boolean
  indicador_ativo?: boolean
  indicador_castrado?: boolean
  microchip_numero?: string
  nascimento_data?: string
  outros_tutores?: ModelsOtherTutor[]
  pedigree_indicador?: boolean
  pedigree_origem_nome?: string
  porte_nome?: string
  qrcode_payload?: string
  raca_nome?: string
  registro_data?: string
  sexo_sigla?: string
  /** "curated" or "self_registered" */
  source?: string
  vermifugacao_data?: string
  vermifugacao_validade_data?: string
}